requirements.txt
paho-mqtt


